requirements.txt
pyyaml==6.0.1
requests==2.28.1